Bug 1481203 - Add new release-mark-as-shipped scopes for Thunderbird r=dustin,tomprince
In Bug 1466627, the scopes for shipit-worker were changed. The original `*:shipit:production` and '*:shipit:staging` were changed to  `*:shipit:server:{staging,production}` (to contrast with the `*:actions:mark-as-{started,shipped}` that were also added there. This adjusts the scopes granted to comm-beta and comm-release to reflect that change.

Additionally, the duplicated stanzas for comm-beta and comm-release were combined.

Differential Revision: https://phabricator.services.mozilla.com/D8592