Bug 1561687 - Move to modern instance types for github workers and stop restricting cores r=tomprince
Move to modern instance types for github workers and stop restricting cores.

For a long time (afaik) we have intended to only have one task running per instance type in github-workers and we had that for 2 of the instance types but not the third. Confusingly, that was also the smallest instance type in terms of memory. Also we had `restrictCPU` set on these which sneakily sets capacity to the number of cores on the instance so those instance types had high capacity as well.

In this change I have set the capacity to 1 and chosen to only use m5d.xlarge instance type as recommended in the bug.

Differential Revision: https://phabricator.services.mozilla.com/D36660