Bug 1564309: Allow people with level-3 access to use interactive tasks; r=dustin
This used to be granted via the roles for level-3 repos (previously the
`moz-tree:level:3` role). Since we don't grant level-3 users the scopes of
level-3 repos anymore, we need to grant this permission directly. Since the
repos *don't* need to access tasks, remove the scope from them at the same
time.

Differential Revision: https://phabricator.services.mozilla.com/D37438