Bug 1618427: Allow actions to be run on cron and release-promotion graphs; r=Callek
Actions can only be run on graphs that have in-tree:hook-action:<hook-id>
scopes for the decision or action task. This works fine on treeherder, which
always looks at the on-push action, but doesn't work on taskcluster, which
looks at the task-group.

Add the scopes to the cron and release-promotion action, to enable actions there too.

Differential Revision: https://phabricator.services.mozilla.com/D64501