Bug 1636297: Increase the concurrency of the hg-push workers; r=Callek
I was investigating the usage of the infra/hg-push workers, and discovered
that the maximum number of concurrent tasks is 4.

Given that the tasks take ~2s to run, we should be able to use a single
instance to handle the load. Trying to run up to 8 concurrent tasks on a single
instance I saw tasks taking up to 8s, which is still acceptable.

Given that we only run a single instance, I limited the available instances to
cheaper non-SSD instances. These tasks do not to much Disk IO.

I also investigated using smaller instances, but discovered significant
docker-worker overhead on small instances; given the already minimal cost of
these workers, I did not think investigating further was worthwhile.

Differential Revision: https://phabricator.services.mozilla.com/D80069