Bug 1661836 - Remove m5ad instances from worker pools. r=releng-reviewers,aki
While c5ad instances are based on Zen 2 Epyc, and are on par with the
equivalent c5d, m5ad instances are based on Zen 1 Epyc and don't keep up
compared to m5d or c5d. For instance, decision tasks running on m5ad
take close to twice the time as on other instance types, while
essentially CPU bound.

Differential Revision: https://phabricator.services.mozilla.com/D88677