No Bug - Remove '4' space yamllint indent rule, r=releng-reviewers,gabriel
The default is that spaces must be consistent, so it'll allow either 2 or 4
spaces as long as they're all the same.

I'm doing this because the yaml files under 'taskcluster' are all using 4 space
indent, but the tool I'm planning to use to format the configuration yaml files
only does 2 space indent.

For the record, 2 space indent is considered the "standard" indentation for
YAML and is what most editors and tools support by default.

Differential Revision: https://phabricator.services.mozilla.com/D145117