No Bug - Avoid using re.escape on trust_domain for in_tree_actions, r=releng-reviewers,aki
This fixes the problem where the `app-services` hooks perpetually show up as
being added in `ci-admin diff`.

The problem occurred because we used `re.escape` on the trust_domain, which
escapes the `-` character even though that character only has meaning inside a
character class (e.g `[0-9]`). There's an old Python issue open for this:
https://bugs.python.org/issue18662

Since the `app-services` trust domain contains a dash, the resource ID
generated for it is called `Hook=project-app\-services`. When `tc-admin`
gathers the "current" resources the TC API (for whatever reason) returns the
hookGroupId as `project-app-services` rather than `project-app\-services`.

The former fails the following check and is considered to be "not managed" and
is therefore ommitted from the diff:
https://github.com/taskcluster/tc-admin/blob/3442cecc49ef89cd80dacf9b16d340f14aff3a68/tcadmin/current/hooks.py#L17

The simplest way to fix this is to avoid calling `re.escape` on the trust
domain here. We don't use any other regex special characters in trust domains
so this shouldn't cause any problems.

Differential Revision: https://phabricator.services.mozilla.com/D147337