Bug 1818075 - Support 'trust_project' routes in the trust-domain-scopes feature, r=releng-reviewers,bhearsum
This splits the `trust-domain-scopes` into three:

1. Scopes that aren't tied to a `trust_project`
2. Scopes for projects without a `trust_project`
3. Scopes for projects with a `trust_project`

This way, if a `trust_project` is specified, the `trust-domain-scopes` feature
will automatically grant caches / routes / secrets that have the
`trust_project` baked in. If `trust_project` is not specified, it will fallback
to the current behaviour where caches / routes / secrets are only based on the
trust domain.

This adds some backwards compatible grants to `mobile-roles` so we avoid
bustage.

Differential Revision: https://phabricator.services.mozilla.com/D170529