No Bug - Rename 'firefox-ci-template' -> 'firefox-ci-playground', r=releng-reviewers,bhearsum
With the upcoming `taskgraph init` feature, there is no longer a need for a
template repo. So instead I've renamed this to a 'playground' and it has been
generated using `taskgraph init`. It will give people a place to try taskgraph
out without Releng needing to set up a new project for each person.

Differential Revision: https://phabricator.services.mozilla.com/D171909