Bug 1842255 - Add 'queueInactvityTimeout' to lifecycle defaults, r=releng-reviewers,bhearsum
This is a new key that was recently added to worker manager. I noticed that the
`ci-admin diff` is starting to get polluted with lots of "removals" of this
value. It seems like all the newer pools are affected, so my theory is that
Taskcluster is implicitly adding this value whenever a pool configuration gets
updated.

By setting this to the same default that Taskcluster sets, we can keep the diff
clean again.

Differential Revision: https://phabricator.services.mozilla.com/D186247