Bug 1879021: centralize ship it scope grants for github projects r=releng-reviewers,jcristau,jlorenzo
We currently grant these both in a fairly generic block, and in some specific project grants. We can get rid of the latter for GitHub projects by improving the existing generic block to put all of them behind a trust domain.

This depends on the adjustment of the scope prefix for Firefox for Android in https://github.com/mozilla-releng/scriptworker-scripts/pull/911.

This change does make some notable changes to scopes, but ultimately they should be harmless:
- We will stop granting ship it scopes to all actions (eg: `Role=repo:github.com/mozilla-extensions/xpi-manifest:action:*`), and instead only grant them to the release promotion action (eg: `Role=repo:github.com/mozilla-extensions/xpi-manifest:action:release-promotion:`). We were also granting scopes to pull requests and pushes in some repositories as well, which is also unnecessary.
- We will stop granting release promotion scopes to adhoc signing repositories. This was added in https://bugzilla.mozilla.org/show_bug.cgi?id=1594621 in preparation for some work that never happened.

Differential Revision: https://phabricator.services.mozilla.com/D200914