Bug 1889647: use user bindable taskcluster proxy port on generic worker machines r=releng-reviewers,jcristau
We recently found that the taskcluster proxy is broken on translations machines due to not being able to bind to port 80. Switching to a port > 1024 fixed it. I'm not certain that all generic worker machines need a user bindable port number, but there should be no harm in using one even where port 80 could be used instead.

Differential Revision: https://phabricator.services.mozilla.com/D206635