Bug 1892526: add generic-worker pools for translations gpu workers r=releng-reviewers,hneiva
We're ready to switch to d2g for the current translations docker worker images. I'm adding these as new pools to ensure we can gracefully switch over in the repository, and have a rollback strategy that doesn't involve ci-config changes. Once things are known to be stable on these new workers, I'll remove the old pools.

Differential Revision: https://phabricator.services.mozilla.com/D208046