Bug 1893250: move github beetmover clients to clients-interpreted.yml r=releng-reviewers,gbrown
The only notable thing here is that some (but not all) beetmover clients need an extra scope to fetch trust-domain-prefixed private artifacts. I _could_ just do this for all of them, but I figured it was best to limit this to what was strictly necessary.

The diff between this and the prior revisions are functionally identical: the only differences are in client descriptions.

Depends on D207294

Differential Revision: https://phabricator.services.mozilla.com/D208536