Bug 1903776: add support for multiple branches per project r=releng-reviewers,ahal
I think this is _fairly_ straightforward in the end: the essence of the change is to add `branches` to the `Project` class. This has implications for all code that deals with branches, default branches, and levels - most notably, there is no such thing as a `level` for a project anymore - it is branch specific. (For now, I am replacing project level with "level of project's default branch" in some places, but even that may be worth reconsidering).

Differential Revision: https://phabricator.services.mozilla.com/D215223