bug 1907217: remove `*` branch from vpn repository; restore it for Maple
Removing the `*` branch from RelEng Github repos worked fine. `mozilla-vpn-client` is the next step for Github, which uses release promotion actions and is higher activity.

On the Mercurial side, we decided to keep the `*` branches since these will be going away sometime next year anyways, so I'm restoring it for `maple`.