bug 1907217: remove * star branches from remaining github repositories
This also adds branches to a couple of repositories where I noticed we were missing one that the build on.

This depends on a handful of PRs to these repos:
https://github.com/mozilla-mobile/reference-browser/pull/2982
https://github.com/mozilla-mobile/firefox-ios/pull/21357
https://github.com/mozilla/code-review/pull/2289
https://github.com/mozilla/code-coverage/pull/2408
https://github.com/mozilla/firefox-translations-training/pull/777

After this change, the only Github repositories with `*` branches will be `staging` repositories, as was discussed in https://github.com/mozilla-releng/fxci-config/pull/18.