add 2tb gpu workers for translations
We're seeing out of disk issues in https://firefox-ci-tc.services.mozilla.com/tasks/DQeRyr1_TjmXhC0Z-5KWWw/runs/1, which is suspected to be an OpusTrainer bug. In the short term, we're going to workaround this by adding 2tb workers. In the medium term we'll fix the root cause and remove these.

Disks are pretty cheap (and certainly a tiny percentage of our translations spend), so while this is not ideal, it's not hugely expensive to do.