feat: add helper to fetch lists of branches from github
This is a basic implementation that should be fine for now. At some point we'll want to convert this to query as an App so that we can support private repositories, but I don't think it's worth the effort at the moment, as it will require additional plumbing for the app secret, installing to a bunch of repos, etc.