feat: generate in tree actions for globbed branches in github repositories
There's a fair amount going on here, but most of it is moving around code (most notably the `branch_name` check in `update_resources` needs to happen before we try to check the level of a branch, as not all branches that exist in Github will be present here). The functional changes here invert our iteration over branches: rather than just assuming all non-globbed ones exist in Github, we fetch the list from Github and check each one to see if it matches one that's configured in projects.yml.