feat: add support for in tree actions from all configured branches (#73)
* feat: make GITHUB_TOKEN accessible to ci-admin github workflows

This secret exists automatically (see https://docs.github.com/en/actions/security-for-github-actions/security-guides/automatic-token-authentication#about-the-github_token-secret). We need it set in a to avoid rate limits when fetching lists of branches for repositories in a subsequent commit.

* feat: add helper to fetch lists of branches from github

This is a basic implementation that should be fine for now. At some point we'll want to convert this to query as an App so that we can support private repositories, but I don't think it's worth the effort at the moment, as it will require additional plumbing for the app secret, installing to a bunch of repos, etc.

* feat: generate in tree actions for globbed branches in github repositories

There's a fair amount going on here, but most of it is moving around code (most notably the `branch_name` check in `update_resources` needs to happen before we try to check the level of a branch, as not all branches that exist in Github will be present here). The functional changes here invert our iteration over branches: rather than just assuming all non-globbed ones exist in Github, we fetch the list from Github and check each one to see if it matches one that's configured in projects.yml.

* fix: add warning when GITHUB_TOKEN is not present in the environment