feat: add support for fetching upstream tasks in find_tasks
There's a lot here, but the essence of this is using `get_ancestors` to find tasks that are upstream of those in `task-graph.json` that we should also rerun in staging. This necessitates some massaging of the task definitions we receive from `get_ancestors` to get them in a similar form, but is other straightforward and unremarkable.

Note that the massaging is done in `util/integration.py` to centralize it, and make the return value consistent regardless of where the tasks came from.