ci: remove 'continue-on-error' from deploy action
It appears that setting this will cause the 'apply' job to report a
successful status even if it failed, which is not what we want.

We want the `notify` job to run regardless of whether its dependencies
were successful or not. From the docs, it looks like setting `if:
always()` should already accomplish this. Therefore I think it's safe to
remove `continue-on-error`.