feat: translations integration tests (#210)
* feat: add support for fetching upstream tasks in find_tasks

There's a lot here, but the essence of this is using `get_ancestors` to find tasks that are upstream of those in `task-graph.json` that we should also rerun in staging. This necessitates some massaging of the task definitions we receive from `get_ancestors` to get them in a similar form, but is other straightforward and unremarkable.

Note that the massaging is done in `util/integration.py` to centralize it, and make the return value consistent regardless of where the tasks came from.

* feat: add support for rewriting dependencies to local versions in integration tests

This ensures that any tasks which will be run in the staging cluster that are upstream of other tasks will be referred to correctly in the downstream tasks.

* fix: don't check excludes against attributes unless there's excludes

`attrmatch` will always return true if `kwargs` is empty; therefore we can't call it unless we have excludes, otherwise we'll reject all tasks.

* feat: add support for optionally mirroring public artifacts in firefoxci-artifact tasks

This work uncovered a small bug in the `firefoxci_artifact` transform where fetches would not be mirrored for tasks without a docker image; this is being fixed here as well.

* feat: fetch translations artifacts to support pipeline integration tests

* feat: add translations pipeline integration tests

* add a schema for the integration test transforms

* set datestamps for all tasks in `find_tasks`

This came up during some additional testing of translations integration tests where the deadline of the `all-pipeline` task was set further in the future than the staging cluster allows. Another commit in this PR already ensures that datestamps are approriate for any ancestor tasks found - there's no harm in doing the same for tasks that come from `task-graph.json`. (Arguably, it's even a feature that we lower the expiration time, as these tasks outlive their usefulness very quickly.)

* remove `REV` from integration test task definitions

This might be the most hacky thing we're doing here so far, but it's necessary for translations integration tests, where cached ancestor tasks will have a `REV` that doesn't match the `REF`, leading to errors such as:
```
[vcs 2024-12-20T16:35:08.992Z] executing ['git', 'config', '--global', '--add', 'safe.directory', '/builds/worker/checkouts/vcs']
[vcs 2024-12-20T16:35:08.994Z] executing ['git', 'clone', 'https://github.com/mozilla-releng/staging-firefox-translations-training', '/builds/worker/checkouts/vcs']
[vcs 2024-12-20T16:35:08.995Z] Cloning into '/builds/worker/checkouts/vcs'...
[vcs 2024-12-20T16:35:10.747Z] executing ['git', 'fetch', '--tags', '--force', 'https://github.com/mozilla-releng/staging-firefox-translations-training', 'refs/heads/main']
[vcs 2024-12-20T16:35:10.949Z] From https://github.com/mozilla-releng/staging-firefox-translations-training
[vcs 2024-12-20T16:35:10.949Z]  * branch            main       -> FETCH_HEAD
[vcs 2024-12-20T16:35:10.955Z] executing ['git', 'fetch', '--no-tags', 'https://github.com/mozilla-releng/staging-firefox-translations-training', 'refs/heads/main']
[vcs 2024-12-20T16:35:11.128Z] From https://github.com/mozilla-releng/staging-firefox-translations-training
[vcs 2024-12-20T16:35:11.128Z]  * branch            main       -> FETCH_HEAD
[vcs 2024-12-20T16:35:11.133Z] executing ['git', 'checkout', '-f', '-B', 'refs/heads/main', '876ad1a2de046271a45c979551c8650e5e1b75a2']
[vcs 2024-12-20T16:35:11.135Z] fatal: reference is not a tree: 876ad1a2de046271a45c979551c8650e5e1b75a2
```

Removing the `REV` will cause the `REF` to be used without issues:
```
[vcs 2024-12-20T18:15:48.936Z] executing ['git', 'config', '--global', '--add', 'safe.directory', '/builds/worker/checkouts/vcs']
[vcs 2024-12-20T18:15:48.938Z] executing ['git', 'clone', 'https://github.com/mozilla-releng/staging-firefox-translations-training', '/builds/worker/checkouts/vcs']
[vcs 2024-12-20T18:15:48.939Z] Cloning into '/builds/worker/checkouts/vcs'...
[vcs 2024-12-20T18:15:50.544Z] executing ['git', 'fetch', '--tags', '--force', 'https://github.com/mozilla-releng/staging-firefox-translations-training', 'refs/heads/main']
[vcs 2024-12-20T18:15:50.758Z] From https://github.com/mozilla-releng/staging-firefox-translations-training
[vcs 2024-12-20T18:15:50.758Z]  * branch            main       -> FETCH_HEAD
[vcs 2024-12-20T18:15:50.763Z] executing ['git', 'fetch', '--no-tags', 'https://github.com/mozilla-releng/staging-firefox-translations-training', 'refs/heads/main']
[vcs 2024-12-20T18:15:50.926Z] From https://github.com/mozilla-releng/staging-firefox-translations-training
[vcs 2024-12-20T18:15:50.926Z]  * branch            main       -> FETCH_HEAD
[vcs 2024-12-20T18:15:50.931Z] executing ['git', 'checkout', '-f', '-B', 'refs/heads/main', 'FETCH_HEAD']
[vcs 2024-12-20T18:15:51.929Z] Switched to a new branch 'refs/heads/main'
```