fix: ignore new and old names for docker image tasks in translations integration tests
These tasks changed names when Translations picked up a new Taskgraph version. At the moment we're in a state where both are present in the most recent cron run (some docker images have been rebuilt, others haven't); we need to keep both in here for awhile to avoid accidentally trying to rebuild them in staging (which doesn't work, and is wasteful).