clients: add non-ldap client for jcristau
When logging in to the UI via auth0 I end up with an auth certificate
larger than 4k (due to the list of ldap and mozillians groups), causing
breakage because it's included in bewit tokens in URLs and the
Authorization header.  Add a static client with star scopes as a
workaround.