ci: fix Gecko os-integration tests
The tasks on the Gecko side were refactored to not be a variant, but I
forgot that there was a filter based on that over here on the
fxci-config side.

I don't think it's important that we filter the "exact" same set of
tasks. This is mainly just to avoid picking up tasks that aren't on the
"happy" path. Builds and tests are mostly all we care about, so start
with that. We can adjust later if more are needed.