feat: don't fail cron runs when pushlog is empty
This seems to be a footgun when setting up a new repository: you can't add it to fxci-config until you push to it, which is both unintuitive, and seems less than ideal (any push you make wouldn't have jobs run, i think?).

Theoretically this could happen if there's an issue on hg.m.o, but I'm not sure that's worth worrying about.