bug 1973252: chore: grant patricia wong access to run release promotion in L1 repositories (#438)
Patricia is working on adding new types of update tests that run as part of release promotion. She needs to be able to test them on Try to keep working on them without running through us.

I don't think we need to bother with Ship It access for this; I plan to show her how to run them through Treeherder directly (in particular, so she can avoid needing new builds each time she pushes to Try).