fix: make translations integration tests work without mirroring fetch tasks
This is a second attempt at a fix for https://bugzilla.mozilla.org/show_bug.cgi?id=1961216, which avoids the [timebomb I called out when it was reviewed](https://github.com/mozilla-releng/fxci-config/pull/361/files#r2070269056).

As it turns out, I already solved the problem of needing fetches from two clusters for toolchain tasks, where we have some that depend on other toolchains + fetches. Backing out the original fix for this, and marking `corpus-clean-parallel-bicleaner-ai` as a task that needs its public upstream artifacts mirrored solves this.

It still feels like there's a bug here in that with the fetch tasks mirrored, we shouldn't need to rely on `mirror-public-fetches` to use the staging cluster version of the tasks. I haven't had time to run that down though, and this should get these tests working again...