feat: support rendering tc.yml with cron hook input
When triggering a cron hook, it's possible to specify a payload at
trigger time. For the "action" type of cron task, it's further possible
to forward this input to the action as action input. But for the
decision cron type, this input is dropped.

This change adds a way for the input to get added to the context used to
render the .taskcluster.yml file, which allows consumers to use it with
the "decision" type of cron.