fix: add 'taskcluster_root_url' to build-decision render context
The Taskcluster Github service renders .taskcluster.yml with a
`taskcluster_root_url` context. We started relying on this in
fxci-config as part of the integration test changes.

However, this broke the `test-build-decision` cron hook, which *doesn't*
render .taskcluster.yml with `taskcluster_root_url`. We could simply
update our `.taskcluster.yml` to check if the context exists before
using it, but IMO we should make the context between the two places
(Github service and build-decision) as close to each other as possible.

So let's ensure we're passing `taskcluster_root_url` as context.