Use the shipit project feature rather than giving scopes manually to shipit
While looking at adding more scopes I got confused as to why we only had
gecko/comm/mobile in the list of scopes added to the shipit clients.
Turns out that those scopes are inferred for other projects using the
`shipit` feature. gecko/comm are the only trust domains that did not
have a single project with the shipit feature, the mobile scopes were
useless. I ended up marking all gecko/comm branches interacting with
shipit with the feature (technically we only need on project per trust
domain/level but I wanted to avoid the confusion).

The only resulting changes are the description of the clients and that
it's going to create a comm-1-dev client.