test: fix 'check_signing_scopes' to only resolve scopes from branches with the requested level
I wanted to set up a repo with:

```
branches:
    - name: main
      level: 3
    - name: project
      level: 1
```

But with the old logic we were appending e.g:

```
assume:repo:github.com/owner/repo:*
```

After expanding, this included both the scopes from the L1 and L3
branches, and the check fails.

While fixing this, I also noticed that for the two org wide projects,
the scope prefix itself ends with a `*`. So I also added a fix for that.