Lift common configuration to worker-defaults for generic-worker/linux-d2g (#612)
This disables native payloads on all d2g pools, enables interactive
tasks on non-trusted ones, and changes the location of generic-worker
caches and downloads to a location that also works on workers with local
storage.

On the other hand, create a new generic-worker/linux-wayland section
that turns off d2g and headlessTasks.