Add an empty task when a github comment doesn't trigger a decision task (#661)
Because this repository has 2 taskcluster instances (staging/prod)
connected to it and when we use a PR comment to trigger an action, we
only want it to happen on staging, the other taskcluster instance will
complain loudly that the comment didn't trigger any action. This is fine
for normal use cases but for ours it's just noisy for no reason.
Instead, this will trigger a built-in/succeed action on github comment
if it didn't match the decision task creation.

I left the checks routes on it so that we could still see that a comment
created a task which would help to debug if ever we see comments not
creating decision tasks (by seeing if it created that skipped one
instead, or did nothing).